Ajustes nos Controllers e Criação de Testes

+++ src/main/java/br/com/fiap/PetPass/controller/EstabelecimentoController.java
@@ -3,6 +3,9 @@ package br.com.fiap.PetPass.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import br.com.fiap.PetPass.dto.EstabelecimentoDTO;
+import br.com.fiap.PetPass.entity.Estabelecimento;
 import br.com.fiap.PetPass.service.EstabelecimentoService;
 
 @RestController
@@ -24,28 +28,52 @@ public class EstabelecimentoController {
 	private EstabelecimentoService service;
 	
 	@GetMapping
-	public List<EstabelecimentoDTO> getAll(){
-		return this.service.findAll();
+	public ResponseEntity<List<EstabelecimentoDTO>> getAll(){
+		List<EstabelecimentoDTO> estabelecimentos = this.service.findAll();
+
+		if(estabelecimentos.isEmpty()) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		return ResponseEntity.ok(estabelecimentos);
 	}
 	
 	@GetMapping("{id}")
-	public EstabelecimentoDTO getById(@PathVariable(value = "id") Long id) {
-		return this.service.findById(id);
+	public ResponseEntity<EstabelecimentoDTO> getById(@PathVariable(value = "id") Long id) {
+		EstabelecimentoDTO estabelecimento = this.service.findById(id);
+		
+		if(estabelecimento != null) {
+			return ResponseEntity.ok(estabelecimento);
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
 	@GetMapping("/documento")
-	public EstabelecimentoDTO getByDocumento(@RequestParam Integer documento) {
-		return this.service.findByDocumento(documento);
+	public ResponseEntity<EstabelecimentoDTO> getByDocumento(@RequestParam Long documento) {
+		EstabelecimentoDTO estabelecimento = this.service.findByDocumento(documento);
+		
+		if(estabelecimento == null) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		return ResponseEntity.ok(estabelecimento);
 	}
 	
 	@GetMapping("/nome")
-	public EstabelecimentoDTO getByNome(@RequestParam String nome) {
-		return this.service.findByNome(nome);
+	public ResponseEntity<EstabelecimentoDTO> getByNome(@RequestParam String nome) {
+		EstabelecimentoDTO estabelecimento = this.service.findByNome(nome);
+		
+		if(estabelecimento == null) {
+			return ResponseEntity.notFound().build();
+		}
+		
+		return ResponseEntity.ok(estabelecimento);
 	}
 	
 	@PostMapping
-	public EstabelecimentoDTO create(@RequestBody EstabelecimentoDTO estabelecimento) {
-		return this.service.create(estabelecimento);
+	public ResponseEntity<EstabelecimentoDTO> create(@RequestBody EstabelecimentoDTO estabelecimento) {
+		EstabelecimentoDTO estabelecimentoDTO = this.service.create(estabelecimento);
+		return ResponseEntity.status(HttpStatus.CREATED).body(estabelecimentoDTO);
 	}
 	
 	@PutMapping
@@ -53,9 +81,16 @@ public class EstabelecimentoController {
 		return this.service.update(estabelecimento);
 	}
 	
-	@DeleteMapping
-	public void delete(@PathVariable Long id) {
+	@DeleteMapping("{id}")
+	public ResponseEntity<?> delete(@PathVariable Long id) {
+		EstabelecimentoDTO estabelecimento = service.findById(id);
+		
+		if(estabelecimento == null) {
+			return ResponseEntity.notFound().build();
+		}
+	
 		this.service.delete(id);
+		return ResponseEntity.ok(id);
 	}
 	
 }
